Use atomic operations when creating and storing a dependency
The map holding the dependencies is now a concurrent map. This ensures
that no value might be erased or wrongly written in case of concurrent
access.

Atomically, the storing operations is done only if there is no previous
mapping. So if two dependencies of same type are created concurrently,
at most two might be created and only one would ever be used (the other
would be thrown). The creation of an extra dependency is not a big issue
and the strong requirement is to never use more than one dependency per
type.