Support live-bidding in Standalone mode (#73)
Going forward, bids will be fetched live from CDB prioritarily, and
fetched from the bid-cache if time-budget is exceeded. This requires
passing the CdbResponseSlot through a callback given the async nature of
the operation.

At first, live-bidding will be enabled for 5% of the users. This will
allow measuring the effectiveness of this strategy and determine the
most optimal time-budget value. To support the A/B test, a new CaC has
been added: `LiveBiddingEnabled`.