Rename BidListener into CdbResponseSlotListener
A listener, that will be part of the public API, will be needed for
supporting live-bidding in InHouse and AppBidding integrations. This
listener will be used by publishers to trigger an action once a bid is
fetched asynchronously.

`BidListener` will be a good candidate for such public API, and to avoid
any confusion the current `BidListener` is renamed into a
`CdbResponseSlotListener`.