Fix InHouse when used with new constructors
There was a condition in the `loadAd` skipping the processing if the ad
unit in the bid token is not the same than the one given in the
constructor.

As the new constructors for InHouse are not taking anymore an ad unit,
this condition was always false, leading to InHouse ads not loading
anymore.