Support dependency provider with lambda capturing "this"
As we're moving forward on Java 8 feature, lambdas will appear in the
dependency provider. Currently the DependenciesAnnotationInjection does
not handled method generated by Android compiler when a lambda is
capturing `this`.

This is fixed by ignoring synthetic methods. Synthetic elements are the
elements added by a Java compiler but without any trace in the source
code. Hence it is legitimate to ignore such methods.