Move check of interstitial availability after the bid
The interstitial format is available only if the
`CriteoInterstitialActivity` activity is registered correctly in the
Android manifest.

This occurred when publishers were using distributed AAR files.
Now that the SDK is distributed only through Maven repositories, such
unavailability should not happen.

This is not totally verified and checking this via remote logs could
determine if it is a useless check.

This commit might make Criteo bids for nothing because the SDK won't be
able to display the ad. Although, `Criteo#loadBid` API should be shared
between InHouse and AppBidding. Therefore anything related to InHouse
should be moved from that call.

JIRA: EE-1224