Add @Internal in BidResponse on consume* methods
The BidResponse class contains public method (`getPrice`) available for
publisher and internal methods used by InHouse.
The visibility is already handled by proguard rules (via @Keep`
annotation) but it is not clear that internal methods are internal.

Now, it is not possible to migrate to Kotlin because Proguard does not
work on it (misconfiguration ?) and even if it was working, it would not
be possible to use the `internal` keywords: Java code cannot access it
because method names are mangled.

Finally the internal is explicit just by using the `@Internal`
annotation instead.

JIRA: EE-1225