Make CriteoBannerView a custom view class
In order to use `CriteoBannerView`, publishers have to instantiate it by
hand and add it to a parent layout programmatically. This is both
inefficient as it creates an unnecessary depth in the view tree, and
tedious.

By turning `CriteoBannerView` into a custom view, publishers will be
able to define it in the layout XML file directly.