Reduce memory retention by nullifying bid listener after consuming it
Options to automatically handle memory retention are:
* (1) Encapsulate listener in a `WeakReference`
* (2) Attach listener to a known lifetime (such as an `Activity`) via
`WeakHashMap`

Option (1) is not possible. Except the `WeakReference`, no one else hold
the listener. Then it is softly reachable and can be GCed.

Option (2) is not possible for two reasons:
- During the Criteo#loadBid, the publisher only gives an ad unit and the
listener to the SDK. So it is impossible to know the context, the
activity.
- `WeakHashMap` solution does not work if listener has a reference to
the key.

JIRA: EE-1297