Simplify loading of interstitials in test app
For Standalone and InHouse, before there was a distinction between:
- valid bid is received
- ad payload is received

To show this distinction, the bid was requested when the activity was
resumed. If valid bid is received, then a button is enabled. Testers had
to click on the button to display the Ad.

This distinction is removed: listener is notified when valid bid is
received AND ad payload is received. The testing process is
simplified to:
- Button is always enabled
- (Manual) Tester clicks on the button
- Bid is requested
- If bid is valid, then ad payload is requested
- If ad payload is received, then interstitial is shown
- (Manual) Tester checks quality of the Ad