Force isolation of tests by waiting for idle state when finished
This also automatically remove registered ActivityLifecycleCallbacks.

This is a necessary fix to solve concurrency issues.
The root issue is that the AppLifecycleUtil registers itself as a
callback to the Application. And this callbacks might be notified even
after the `waitForIdleState` below, so after the end of a test while a
new test is running.

When the AppLifecycleUtil is notified, it starts an async task, creates
logger, fetches network, ...
Then this yields to concurrency issues such as:
- new test: boostrap new dependency provider (1)
- new test: start to stub the dependencies: doReturn(foo).when(dependencyProvider)
*context switch*
- old test: AppLifecycleUtil is notified and start an async task
- old test: a new SafeRunnable is created, creating a new logger via:
`DependencyProvider.getInstance().provideLoggerFactory()...`

The issue start here: the DependencyProvider.getInstance() returns (1),
the dependency provider of the new test that was being stubbed. And the
stub is finally:
```
doReturn(foo).when(dependencyProvider).provideLoggerFactory()
```

Some exceptions might appear later: ClassCastException,
NullPointerException or WrongTypeOfReturnValue.
It appears on another thread making the debugging a nightmare.