Disable analytics sent to Google when running tests
The Android JUnit runner is sniffing some data when we run our tests.
The worse part is that sometimes the network connection is bad (I can't
tell if it was an issue client side or server side) and, after tests are
really finished, the tests process takes many seconds to stop.