Provide test components via dependency provider too
Few components are available during test and might profit from the
dependency system.
There internal components of the MockedDependenciesRule are provided:
Logger and CdbMock.

Other components might follow the same logic:
- WebViewLookup
- WebViewClicker
- InstrumentationUtil
- CriteoUtil

The idea is to centralized dependencies even for tests, so all
components are shareable. For instance, WebViewLookup needs a context
and is getting it statically on its own.