Fix functional test on AppEvent
This test is broken because of 2aa15017: AsyncTack of AppEvent is
replaced by Executor+Runnable. With AsyncTask, only a single GUM request
can be executed at the same time. With Executor, many GUM requests can
be done in parallel.

With new solution, it appears that two events are sent when starting a
new activity: Launch then Active events.

The former approach is tedious and not deterministic because if network
is slow, then only the "Launch" event is sent, if network is fast, then
both events are sent.

So test is updated to accept the two events. The @FlakyTest is also
removed on this test:
- the new solution is deterministic
- recent work was done to isolate tests, particularly on AppEvents