Generify SendingQueue implementation and creation
It is generify to both CSM and Logger can use this concept.

The dependency provider does not support well generic class because
dependencies are stored by class, and class does not keep the generic
information. So marker classes are used to statically identify the CSM
sending queue and the Logger one.

Tests are generify too, but Metric objects are kept as payload to not
totally rewrite the tests. This is not an issue because metrics are just
there as dummy payload and could be anything (log records, string, ...).
Moreover, functional tests are there to guarantee the integration.

JIRA: EE-1347