Inject the remote handler into the logger
The remote handler is ready to be injected. At this point it will do
nothing except filling its sending queue (which is bounded), although it
is safe to have it.

Logger get handlers via lazy providers. This is to allow any
dependencies in the handlers without worrying about circular
dependencies. For instance, remote handler needs the deviceId but
component fetching it needs a logger.

As Logger lazily gets its dependencies, then the MockedDependenciesRule
don't need to handle that itself and the SpyLoggerResource can be
cleaned. The MockedDependenciesRuleLoggingTest remains to assert that
there is no regressions.

JIRA: EE-1347