Protect infinite recursion from handlers that use logger
The `AdvertiserUtil` can log a failure when getting the GAID throws an
exception.
The `IntegrationRegistry` class can log a failure when shared preference
contains an ill-formed integration name.

As they are used by the RemoteHandler to get context of remote logs, an
infinite recursive loop might be initiated. The logger should then
detect it and stop it.

Such errors can be very uncommon but if they are occurring once, they
might lock the SDK on the device.

JIRA: EE-1347