Make AdvertisingInfo callable on any thread
The Google API is throwing an exception when called on the main thread:
```
java.lang.IllegalStateException: Calling this from your main thread can lead to deadlock
```

The GAID is fetched in 4 cases:
- when creating the CDB request
- when creating the Bearcat request
- when creating the remote config request
- when creating the remote logs request

The 3 first cases are done on worker thread and current solution is ok.
The remote log case can be done on any thread, including the main one
which is problematic.

To solve this issue we can either:
1. run remote log handler/logger in a (dedicated?) worker thread
2. refactor AdvertisingInfo class to have an async API (listener/future)
3. make AdvertisingInfo callable from any thread

# Proposition 1
Pros:
- slow logger cannot impact business
- no change needed on AdvertisingInfo class
Cons:
- threadId field in logcat or remote log became useless
- impossible or really hard to detect recursive logs and repercussion is
disastrous: instead of 1 stack overflow on 1 thread, the SDK can be a
fork bomb taking 100% of CPU and/or RAM

# Proposition 2
Pros:
- make no assumption
Cons:
- huge refactoring needed
- async API are contaminant: all call stack should be refactored

# Proposition 3
Pros:
- only local change required in AdvertisingInfo class
- no change for caller
Cons:
- assume that GAID is constant
- GAID is available for main thread only after a first call

This commit is implementing the proposition 3.

About the assumption that GAID is constant: to limit the GAID, user
should go deeply in device parameter. When user come back to the
application, either the Android killed the app and the SDK is reinit
again, either the app is still alive, then assumption is false.

About the fact that GAID is available for main thread only after a first
call, this is fixed by prefetching it during SDK init.