Fix functional test detecting InHouse profile on native
This test was working before the AdvertiserInfo change just by luck.
Apparently, when getting the GAID, the Google SDK needs to start from a
worker thread but seems to do something on the main thread.
At the same time, the InHouse detection is in the bid response listener,
which is called on main thread. Also, because cache bidding is still the
default mode the bid manager is fetching, in parallel a new bid while
notifying the bid response listener.

With a lot of luck, a good alignment of the threads and the
planets+Pluto, then the InHouse profile ID is detected before the second
CDB call is made.

With the previous arrangement, it was weirdly deterministic but it is
certainly not a stable arrangement and it is then stabilized.