Show millis of client date time when generating remote logs
Remote logs are bulk. So each chunk has the same server time. To be able
to reorder the logs, the client date time is included in the message.
Before, time was printed only down to seconds which is a too big unit of
time. So it was almost impossible to use it.
This commit shows the milliseconds too, so one can distinguish the real
log order.

Maybe for later:
Even with millis, logs might not be fully reorder. An atomic log IDs
could be used instead.