Inject deviceId when missing in persisted log records
Such situation might happen when the log record was created on the main
thread and that the deviceId was not ready yet: for instance for logs
emitted during initialization of the SDK.

The deviceId is assumed to be a constant. Hence, it is not an issue to
set it later. Although, this is quite a workaround based on an
assumption. Hence, the deviceId is kept if it is already set.

JIRA: EE-1411