Fix infinite loop of NoSuchMethodException
The `TapeSendingQueue` class accesses private members of a the
com.squareup.tape dependency:
- `FileObjectQueue#queueFile`
- `QueueFile#usedBytes()`

This privates members are used to determine the real size, in bytes, of
the queue. Then the queue can be capped to avoid consuming too much
memory (disk).

When using this SDK, publishers can apply minification. Since AGP 3.4.0,
R8 is enabled by default. Before Proguard was used. It is still possible
to used Proguard by applying `android.enableRB=false` in the
`gradle.properties`, but this option is deprecated since AGP 3.6.0.

With R8 minification, there is no error to signal.
With Proguard minification and optimization enabled, it is possible to
produce an infinite loop of `NoSuchMethodException` logs.

The explanation is that, with the default Android optimization rules
(proguard-android-optimize.txt), both `method/inlining/short` and
`method/inlining/unique` optimization rules are enabled. (see
https://www.guardsquare.com/en/products/proguard/manual/usage/optimizations).
Such optimization can inline methods at compile time rather than at
runtime via a JIT.

Now, as the `QueueFile#usedBytes()` is private and only used by another
method, it is inlined by Proguard. Hence the `NoSuchMethodException`
when trying to reach it.

The infinite loop of exception comes from the fact that the remote
logger use the `TapeSendingQueue` as a buffer. The exception is thrown
every time, it is logged, given to the remote logger, which use again
the `TapeSendingQueue`, ... indefinitely.

The `NoSuchMethodException` is fixed by indicating to Proguard to keep
elements reached by reflection. The infinite loop is therefore fixed
too, but a security is added to prevent infinite loop on worker threads.

JIRA: DPP-3238