Add module dedicated to test SDK with AdMob before v20
A dedicated module is used because Gradle does not handle having
multiple versions of the same library during the compilation. Also, even
if it was handled, Java does not handle collision of names (because of
two different versions) in the same classloader.

Moreover, once the support of AdMob 19 is over, this whole module can be
safely deleted.

JIRA: DPP-3348