🩹 Catch OOM when checking if queue file is not corrupted
When the queue file gets corrupted, it can write crazy sizes such as
negative ones or huge ones (such as 800MB).

Even if it is not a good practice to catch an error, in this case we can
recover from it by erasing the queue file and creating a new one.

This should fix the symptoms of issue #228, although, the origin of
corruption still need to be determined.