🔊 Throw/Log failures when fetching GAID
In debug (during tests), the code will rethrow the failure. In prod, it
will be logged and send along the remote logger handler.

Fetching the GAID is primordial to bid, and CI shows that sometimes, it
might fail. Hence it can be considered as a proper error and not as a
debug.