✅ Move Session test in JVM tests and make it use the test rule
By using the mocked dependencies rule, this test is able to run on the
JVM and all parallel tasks are automatically caught and wait on so that
the test is isolated from others.

Before, without the rule, the test was initializing the SDK several
times which spawned multiple async tasks that might have side effects on
other tests.