✅ Isolate test by using mocked dependency rule to tear down state
By setting the dependency provider to null, all async tasks are still
live. And newly created dependencies are no more managed by the mocked
dependency rule. So new async tasks are not waited properly at the end
of the test.

This means that the test was previously still living while other tests
were executed.

By using the rule to tear down the in memory state, then all async tasks
are waited on properly and all new dependencies are still managed and
injected by the rule.