👷 Disable coverage check on test-utils module
The test-utils module contains all the shared logic to mock/inject/help
for testing the SDK and even the mediation adapters.

Then it is implicitly tested by the PublisherSDK tests. Although, we
could argue that it might need some dedicated tests, for the moment it
is not and it is mainly a pain during some code reviews.

Hence until code coverage from PublisherSDK tests is not counted, then
checks are disabled.

JIRA: EE-1370