✅ Fix flaky test on live bidding
Those tests were observed to be very flaky. Indeed, it used a real clock
to simulate the live bidding time budget being expired or respected.
This involves 2 executors (one async and one scheduled) which add a lot
of imprecision and difficulties when mocking.

To stabilize the tests, instead of mocking the time budget, the
scheduling is simply mocked to directly call or never call the listener
then.