🩹 Fix flaky test
Here is what happen when requesting a bid for App Bidding (with cache
bidding, it's quite the same for live bidding too):
- get bid from cache (assuming there is one)
- async on the main thread:
  - call the bid listener with bid
  - App bidding handler is called
  - (1) App bidding handler declare that the integration is app bidding
- async in worker thread:
  - (2) prepare a CDB call with the detected integration
  - fetch CDB call and store the response

Step (1) and (2) are running concurrently, hence the order is not
guaranteed.

The fix is to wait for idle state and trigger a new bid request. Hence
we're sure that the (1) of the first bid request is done before the (2)
of the second one, so that at least the second request is using the good
integration.

JIRA: DPP-3624