🤡 Disable detection of mediation adapters by default
The integration registry actively detect if mediation adapters classes
are present in the classpath. If present, then it is assumed that the
integration is the detected adapter. If both are presents, then the
fallback integration is used.

In the case of the test app, both adapters are present. So InHouse,
Standalone and AppBidding are never detected and even adapters are not
detected as the SDK is using the fallback integration.

This is tedious when testing and error prone, as all should think about
this. So, at startup, integration detector is mocked to ignore the
adapters. Also, toast are displayed when a new integration is declared
so that testers will at least be notified accordingly.

It is still possible to use the previous activity to disable this mock
or mock even more.