🐛 Fix rewarded video showcase with GAM
Strangely, GAM throw an error when the VAST XML is not indented as it
expects. This is mainly observable on the companion ad XML snippet: the
previous code yield an error, but just indenting a little bit more the
<CompanionAds> tags is enough to make it works. Also, if all indentation
and newlines are removed, then its working too.

This is how this is fixed: all lines of the XML is trimmed and newlines
are removed at runtime.