♻️ Cleanup a lot of deprecated code by encapsulating shared pref
The `PreferenceManager` class is deprecated and require to add a new
androix library. To prevent adding a new 3P library, but to limit the
deprecation, then the lifecycle of the `SharedPreferences` is now
managed by the `DependencyProvider`.

Also, this means that all persisted side effects are cleaned up
automatically when using the `MockedDependencyProviderRule`.