Remove flaky and tedious test
When calling Criteo.init(), a prefetch request is triggered async.
This test is assuming that the setUsPrivacyOptOut call is done before
the execution of the prefetch. This is not always the case if the
testing thread is preempted and that the prefetch thread get running.
Not sure if this test is really useful then. The scenario is an edge
case and the important is already checked by the test below:
- Given setUsPrivacyOptOut is called
- And new CDB calls are made
- Then the CDB calls are using the given value