🔥 Remove useless test
This test was initially written to check that after 7s, the interstitial
activity is closed automatically. Then it was updated to check that
after 7s, the activity is still there. Then the 7s was reduced to 2s
because 7s is big to wait.
Finally I remove this test because waiting for 2s to check that nothing
happen is quite useless.

But moreover this test make the emulator's CI crash randomly.