♻️ Add a notion of SDK service lifecycle that get initialized at init
The purpose is to move out the initialization code in the CriteoInternal
class or in other classes.

For instance, previously, the config update was done within the prefetch
method of the BidManager. Hopefully tests caught that removing the
prefetch also removed the initial config call. But to ease the
development and the maintenance, it is preferable to centralize such
initialization in a dedicated manager.